requirements.txt
cassandra-driver==3.6.0
Django==1.10.2
django-cassandra-engine==0.11.1
djangorestframework==3.4.7
futures==3.0.5
psycopg2==2.6.2
six==1.10.0
